requirements.txt
redis
flask
pandas
requests
dataclasses_json
PyExecjs
sqlalchemy
pymysql
JPype1
pycryptodome
lxml
requests_toolbelt
phone_gen
httpx
ccard
xmltodict
pycurl
py_mini_racer
marshmallow
python-dateutil
urllib3
beautifulsoup4
